Support for @file:my.file in Command Line Parameters
This patch adds support for @file:my.file in Command Line Parameters. The parameter is replaced by the file contents of the requested file.
By design: only supported in ARGV, not in KEYS. Can be repeated for multiple files, and combined with normal parameters.
Needed when passing binary data (like MSGPACK), or formatted human-readable JSON, to the Redis Lua debug session.